Move versions classes to own package
