+++ src/graphdatascience/__init__.py
@@ -1,8 +1,8 @@
-from .graph.graph_api import Graph
-from .graph_data_science import GraphDataScience
-from .server_version.server_version import ServerVersion
-from .session import GdsSessions
-from .version import __version__
+from graphdatascience.graph.graph_api import Graph
+from graphdatascience.graph_data_science import GraphDataScience
+from graphdatascience.session import GdsSessions
+from graphdatascience.version import __version__
+from graphdatascience.versions import ServerVersion
 
 __all__ = [
     "__version__",
