+++ src/graphdatascience/query_runner/query_runner.py
@@ -6,7 +6,7 @@ from pandas import DataFrame
 from graphdatascience.call_parameters import CallParameters
 from graphdatascience.query_runner.query_mode import QueryMode
 from graphdatascience.query_runner.query_type import QueryType
-from graphdatascience.server_version.server_version import ServerVersion
+from graphdatascience.versions import ServerVersion
 
 
 class QueryRunner(ABC):
